Refactor layout in Home and Sidebar components, update Navbar padding, and implement Profile component with API integration
